Improve video and gif playback
This commit solves the problem with losing audio focus when
playing GIFs. Twitter converts .gif animations to .mp4 videos to
save bandwidth and make manipulation easier.
Previously, playing GIFs in VideoView stole the audio focus from the
audio manager. With this commit, GIFs are played in TextureView.
It also solves the VideoView flickering when loading the video.